Add AutoCompleteManager and remove autocompletetion from chatbar
