+++ Sources/Views/ALKChatBar.swift
@@ -25,9 +25,9 @@ public struct AutoCompleteItem {
     }
 }
 
-public protocol AutoCompletionDelegate: AnyObject {
-    func didMatch(prefix: String, message: String)
-}
+//public protocol AutoCompletionDelegate: AnyObject {
+//    func didMatch(prefix: String, message: String)
+//}
 
 // swiftlint:disable:next type_body_length
 open class ALKChatBar: UIView, Localizable {
@@ -38,7 +38,13 @@ open class ALKChatBar: UIView, Localizable {
     }
 
     public var isMicButtonHidden: Bool!
-    public weak var autocompletionDelegate: AutoCompletionDelegate?
+//    public weak var autocompletionDelegate: AutoCompletionDelegate?
+
+    // TODO: not sure chat bar should be aware of
+    // autocompletemanager, right now adding here
+    // as when textview's delegate gets triggered it will call
+    // manager's methods
+    public var autoCompleteManager: AutoCompleteManager!
 
     public enum ButtonMode {
         case send
@@ -249,21 +255,21 @@ open class ALKChatBar: UIView, Localizable {
 
     fileprivate var textViewTrailingWithSend: NSLayoutConstraint?
     fileprivate var textViewTrailingWithMic: NSLayoutConstraint?
-    fileprivate var autoCompletionViewHeightConstraint: NSLayoutConstraint?
+//    fileprivate var autoCompletionViewHeightConstraint: NSLayoutConstraint?
 
-    public var autoCompletionItems = [AutoCompleteItem]()
-    var filteredAutocompletionItems = [AutoCompleteItem]()
+//    public var autoCompletionItems = [AutoCompleteItem]()
+//    var filteredAutocompletionItems = [AutoCompleteItem]()
 
-    private var autocompletionPrefixes: Set<String> = []
-    private var autocompletionPrefixAttributes: [String: [NSAttributedString.Key: Any]] = [:]
+//    private var autocompletionPrefixes: Set<String> = []
+//    private var autocompletionPrefixAttributes: [String: [NSAttributedString.Key: Any]] = [:]
 
     // Prefix and selected item pair
-    typealias Selection = (
-        prefix: String,
-        range: NSRange,
-        word: String
-    )
-    var selection: Selection?
+//    typealias Selection = (
+//        prefix: String,
+//        range: NSRange,
+//        word: String
+//    )
+//    var selection: Selection?
 
     private enum ConstraintIdentifier: String {
         case mediaBackgroudViewHeight
@@ -335,18 +341,18 @@ open class ALKChatBar: UIView, Localizable {
         updateMediaViewVisibility()
     }
 
-    func setupAutoCompletion(_ tableview: UITableView) {
-        autocompletionView = tableview
-        autocompletionView.dataSource = self
-        autocompletionView.delegate = self
-        autoCompletionViewHeightConstraint = autocompletionView.heightAnchor.constraint(equalToConstant: 0)
-        autoCompletionViewHeightConstraint?.isActive = true
-    }
-
-    func registerPrefix(prefix: String, attributes: [NSAttributedString.Key: Any]) {
-        autocompletionPrefixes.insert(prefix)
-        autocompletionPrefixAttributes[prefix] = attributes
-    }
+//    func setupAutoCompletion(_ tableview: UITableView) {
+//        autocompletionView = tableview
+//        autocompletionView.dataSource = self
+//        autocompletionView.delegate = self
+//        autoCompletionViewHeightConstraint = autocompletionView.heightAnchor.constraint(equalToConstant: 0)
+//        autoCompletionViewHeightConstraint?.isActive = true
+//    }
+//
+//    func registerPrefix(prefix: String, attributes: [NSAttributedString.Key: Any]) {
+//        autocompletionPrefixes.insert(prefix)
+//        autocompletionPrefixAttributes[prefix] = attributes
+//    }
 
     func setComingSoonDelegate(delegate: UIView) {
         comingSoonDelegate = delegate
@@ -357,7 +363,7 @@ open class ALKChatBar: UIView, Localizable {
         clearTextInTextView()
         textView.attributedText = nil
         toggleKeyboardType(textView: textView)
-        hideAutoCompletionView()
+//        hideAutoCompletionView()
     }
 
     func hideMicButton() {
@@ -706,55 +712,63 @@ open class ALKChatBar: UIView, Localizable {
 
 extension ALKChatBar: UITextViewDelegate {
     public func textView(_ textView: UITextView, shouldChangeTextIn range: NSRange, replacementText string: String) -> Bool {
-        guard var text = textView.text as NSString? else {
-            return true
-        }
-
-        // check if deleting an autocomplete item, if yes then
-        // remove full item in one go and clear the attributes
-        //
-        // range.length == 1: Remove single character
-        // range.lowerBound < textView.selectedRange.lowerBound: Ignore trying to delete
-        //      the substring if the user is already doing so
-        if range.length == 1, range.lowerBound < textView.selectedRange.lowerBound {
-            // Backspace/removing text
-            let attribute = textView.attributedText
-                .attributes(at: range.lowerBound, longestEffectiveRange: nil, in: range)
-                .filter { $0.key == AutoCompleteItem.attributesKey }
-
-            if let isAutocomplete = attribute[AutoCompleteItem.attributesKey] as? String, !isAutocomplete.isEmpty {
-                // Remove the autocompleted substring
-                let lowerRange = NSRange(location: 0, length: range.location + 1)
-                textView.attributedText.enumerateAttribute(AutoCompleteItem.attributesKey, in: lowerRange, options: .reverse, using: { _, range, stop in
-
-                    // Only delete the first found range
-                    defer { stop.pointee = true }
-
-                    let emptyString = NSAttributedString(string: "", attributes: textView.typingAttributes)
-                    textView.attributedText = textView.attributedText.replacingCharacters(in: range, with: emptyString)
-                    textView.selectedRange = NSRange(location: range.location, length: 0)
-                })
-            }
-        }
-
-        text = text.replacingCharacters(in: range, with: string) as NSString
-        updateTextViewHeight(textView: textView, text: text as String)
-        return true
+        return autoCompleteManager.textView(textView, shouldChangeTextIn: range, replacementText: string)
+//        autoCompleteManager.should
+//        guard var text = textView.text as NSString? else {
+//            return true
+//        }
+//
+//        // check if deleting an autocomplete item, if yes then
+//        // remove full item in one go and clear the attributes
+//        //
+//        // range.length == 1: Remove single character
+//        // range.lowerBound < textView.selectedRange.lowerBound: Ignore trying to delete
+//        //      the substring if the user is already doing so
+//        if range.length == 1, range.lowerBound < textView.selectedRange.lowerBound {
+//            // Backspace/removing text
+//            let attribute = textView.attributedText
+//                .attributes(at: range.lowerBound, longestEffectiveRange: nil, in: range)
+//                .filter { $0.key == AutoCompleteItem.attributesKey }
+//
+//            if let isAutocomplete = attribute[AutoCompleteItem.attributesKey] as? String, !isAutocomplete.isEmpty {
+//                // Remove the autocompleted substring
+//                let lowerRange = NSRange(location: 0, length: range.location + 1)
+//                textView.attributedText.enumerateAttribute(AutoCompleteItem.attributesKey, in: lowerRange, options: .reverse, using: { _, range, stop in
+//
+//                    // Only delete the first found range
+//                    defer { stop.pointee = true }
+//
+//                    let emptyString = NSAttributedString(string: "", attributes: textView.typingAttributes)
+//                    textView.attributedText = textView.attributedText.replacingCharacters(in: range, with: emptyString)
+//                    textView.selectedRange = NSRange(location: range.location, length: 0)
+//                })
+//            }
+//        }
+//
+//        text = text.replacingCharacters(in: range, with: string) as NSString
+//        updateTextViewHeight(textView: textView, text: text as String)
+//        return true
     }
 
     public func textViewDidChangeSelection(_ textView: UITextView) {
-        guard let result = textView.find(prefixes: autocompletionPrefixes) else {
-            hideAutoCompletionView()
-            return
-        }
-
-        selection = (result.prefix, result.range, String(result.word.dropFirst(result.prefix.count)))
-        // Call delegate and get items
-        autocompletionDelegate?.didMatch(prefix: result.prefix, message: String(result.word.dropFirst(result.prefix.count)))
+        autoCompleteManager.textViewDidChangeSelection(textView)
+//        guard let result = textView.find(prefixes: autocompletionPrefixes) else {
+//            hideAutoCompletionView()
+//            return
+//        }
+//
+//        selection = (result.prefix, result.range, String(result.word.dropFirst(result.prefix.count)))
+//        // Call delegate and get items
+//        autocompletionDelegate?.didMatch(prefix: result.prefix, message: String(result.word.dropFirst(result.prefix.count)))
     }
 
     public func textViewDidChange(_ textView: UITextView) {
         textView.typingAttributes = defaultTextAttributes
+        // TODO: post notification or create multicast delegate
+        // pefer multicast delegate but notification posting is easy
+        // and we can check for equality
+        // https://github.com/nathantannar4/InputBarAccessoryView/search?q=UITextView.textDidChangeNotification&unscoped_q=UITextView.textDidChangeNotification
+        // and messaviewcontroller for multiple delegates
         if textView.text.isEmpty {
             clearTextInTextView()
         } else {
@@ -830,17 +844,17 @@ extension ALKChatBar: UITextViewDelegate {
         autocompletionView.reloadData()
     }
 
-    func showAutoCompletionView() {
-        let contentHeight = autocompletionView.contentSize.height
+//    func showAutoCompletionView() {
+//        let contentHeight = autocompletionView.contentSize.height
+//
+//        let bottomPadding: CGFloat = contentHeight > 0 ? 25 : 0
+//        let maxheight: CGFloat = 200
+//        autoCompletionViewHeightConstraint?.constant = contentHeight < maxheight ? contentHeight + bottomPadding : maxheight
+//    }
 
-        let bottomPadding: CGFloat = contentHeight > 0 ? 25 : 0
-        let maxheight: CGFloat = 200
-        autoCompletionViewHeightConstraint?.constant = contentHeight < maxheight ? contentHeight + bottomPadding : maxheight
-    }
-
-    func hideAutoCompletionView() {
-        autoCompletionViewHeightConstraint?.constant = 0
-    }
+//    func hideAutoCompletionView() {
+//        autoCompletionViewHeightConstraint?.constant = 0
+//    }
 
     func textStartsWithPrefix(_ text: String, prefix: String) -> Bool {
         guard !prefix.isEmpty, text.starts(with: prefix) else { return false }
@@ -848,33 +862,33 @@ extension ALKChatBar: UITextViewDelegate {
         return true
     }
 
-    func insert(item: AutoCompleteItem, at insertionRange: NSRange, replace selection: Selection) {
-        let defaultAttributes = textView.typingAttributes
-        var newAttributes = defaultAttributes
-        if let prefixAttributes = autocompletionPrefixAttributes[selection.prefix] {
-            // pass prefix attributes for the range and override old value if present
-            newAttributes.merge(prefixAttributes) { $1 }
-        }
-        // prefix to identify which autocomplete is present
-        newAttributes[AutoCompleteItem.attributesKey] = selection.prefix + item.key
-
-        let insertionItemString = NSAttributedString(
-            string: selection.prefix + item.content,
-            attributes: newAttributes
-        )
-
-        let newAttributedText = textView.attributedText.replacingCharacters(
-            in: insertionRange,
-            with: insertionItemString
-        )
-        newAttributedText.append(NSAttributedString(string: " ", attributes: defaultAttributes))
-
-        // If we replace the text here then it resizes the textview incorrectly.
-        // That's why first resetting the text and then inserting the item content.
-        // Also, to prevent keyboard autocorrect from cloberring the insert.
-        textView.attributedText = NSAttributedString()
-        textView.attributedText = newAttributedText
-    }
+//    func insert(item: AutoCompleteItem, at insertionRange: NSRange, replace selection: Selection) {
+//        let defaultAttributes = textView.typingAttributes
+//        var newAttributes = defaultAttributes
+//        if let prefixAttributes = autocompletionPrefixAttributes[selection.prefix] {
+//            // pass prefix attributes for the range and override old value if present
+//            newAttributes.merge(prefixAttributes) { $1 }
+//        }
+//        // prefix to identify which autocomplete is present
+//        newAttributes[AutoCompleteItem.attributesKey] = selection.prefix + item.key
+//
+//        let insertionItemString = NSAttributedString(
+//            string: selection.prefix + item.content,
+//            attributes: newAttributes
+//        )
+//
+//        let newAttributedText = textView.attributedText.replacingCharacters(
+//            in: insertionRange,
+//            with: insertionItemString
+//        )
+//        newAttributedText.append(NSAttributedString(string: " ", attributes: defaultAttributes))
+//
+//        // If we replace the text here then it resizes the textview incorrectly.
+//        // That's why first resetting the text and then inserting the item content.
+//        // Also, to prevent keyboard autocorrect from cloberring the insert.
+//        textView.attributedText = NSAttributedString()
+//        textView.attributedText = newAttributedText
+//    }
 }
 
 extension ALKChatBar: ALKAudioRecorderProtocol {
